refactor: add Auth file and move its content into Auth.jsx along with styles
